Procesos de primer entrega finalizados

+++ src/main/java/uniquindio/product/dto/pedido/MostrarDetallePedidoDTO.java
@@ -2,12 +2,14 @@ package uniquindio.product.dto.pedido;
 
 import uniquindio.product.model.enums.TipoProducto;
 
+import java.math.BigDecimal;
+
 public record MostrarDetallePedidoDTO(
         String idProducto,
         String nombreProducto,
         TipoProducto tipoProducto,
         String imagenProducto,
-        Double precioUnitario,
-        int cantidad,
-        Double subtotal
+        BigDecimal precioUnitario,
+        Integer cantidad,
+        BigDecimal subtotal
 ) {}
